chore: Upgrade Feather Wiki empties to version 1.8.2
Commit created with `VER=1.8.2 make feather-update`

I modified the bird name manually.

Note: I'm not updating the "featherx" empty. Reasoning: It's
designated as experimental, and currently it requires a custom
build. IIUC if we wait for a future release of Feather Wiki (1.9
perhaps?) then the round-trip bare bones saving will be supported
and it won't require a custom build. So let's wait for that.