Add commentary on bootstrap-email sass precompile
I was using a Makefile target to precompile the bootstrap-email
styles so the many deprecation warnings didn't clutter up the test
suite output after updating the dependencies. Let's explain that a
little, since I don't want it to seem like it's important.

I did consider removing it, but it's fairly low maintenance and
harmless, so decided to keep it for now anyway.

The SASS warnings were fixed here IIUC:
https://github.com/bootstrap-email/bootstrap-email/pull/282