Add make target to show stripe api version
I've found that `stripe listen` uses the default api version
associated with stripe account. There's some more work to do to make
it use the right one, but this is useful so we know what version we
need.

Note that it does have a --latest flag, but right now I don't think
the pay and stripe dependencies are up to the latest.