Use --latest with stripe listen for local dev/test
I'm not 100% clear on how all stripe API versioning works, but with
this flag I was able to successfully test a subscribe and
unsubscribe locally. Without this it reports a quite old API version
and I see some errors processing the webhooks.

For the record:
Without --latest I get this:
  Ready! You are using Stripe API Version [2020-08-27].

With --latest I get this:
  Ready! You are using Stripe API Version [2025-11-17.clover].

Yet `make strip-api-version` returns this:
  2025-08-27.basil

And in the Stripe UI I see this for my webhook:
  API version 2022-11-15

(Possibly Stripe API and the Stripe Webhooks API are different
things with different versions, but I'm really not sure..)