Restart rails job handler after migration also
The pay gem migrations ran on deploy, but because the job handler
didn't get a restart it was unaware of the new schema and hence
started producing errors, specifically this one:

  undefined method 'api_record' for an instance of Pay::Customer

Restarting the jobs container fixed it in this case. This ansible
change should hopefully fix it when/if there's a next time.