Use a var for docker push location
The default value in the Makefile is still hard-coded, but in the
GitHub workflow we're now using a repo var.

Note that it's still hard coded in the docker-compose files, but
that's okay for now I guess.

Additionally, add docker.io explicitly, so we could more easily move
to another registry some day.